backend/requirements.txt
flask==2.0.2
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
faker==9.8.2
sqlalchemy-serializer==1.3.1
flask-restful==0.3.9
flask-cors==3.0.10
